Moved scripts out from link folder
