+++ scripts/tools/linkToolsAndroid.js
